\name{real.roots}
\alias{function_real_roots}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Real Roots
%%  Determines the roots of quadratic equation %%
}
\description{
In algebra, a real root is a solution to a particular equation. The term real root means that this solution is a number that can be whole, positive, negative, rational, or irrational. While numbers like pi and the square root of two are irrational numbers, rational numbers are zero, whole numbers, fractions and decimals.
}
\usage{
real.roots(x, y, z)
}

\author{
Thakur Dev Pandey
}

\examples{
## real.roots(1, 5, 6)
[1] -3 -2

}
